requirements.txt
Flask
chromadb
docx
google-generativeai
langchain
pandas
pypdf
python-pptx
